fix ZapReceipt.isValid
