+++ packages/ndk/lib/domain_layer/usecases/zaps/Invoice_response.dart
@@ -1,0 +1,9 @@
+/// invoice response
+class InvoiceResponse {
+  String invoice;
+  int amountSats;
+  String? nostrPubkey;
+
+  /// .
+  InvoiceResponse({required this.invoice, this.nostrPubkey, required this.amountSats});
+}
