+++ packages/ndk/lib/domain_layer/usecases/zaps/zap_receipt.dart
@@ -2,7 +2,6 @@ import 'dart:convert';
 
 import 'package:ndk/domain_layer/entities/nip_01_event.dart';
 import 'package:ndk/domain_layer/usecases/lnurl/lnurl.dart';
-import 'package:ndk/domain_layer/usecases/zaps/zap_request.dart';
 
 import '../../../shared/logger/logger.dart';
 
@@ -19,9 +18,11 @@ class ZapReceipt {
   String? comment;
   String? sender;
   String? anon;
+  String? lnurl;
 
   ZapReceipt.fromEvent(Nip01Event event) {
     String? zapRequestJson;
+    pubKey = event.pubKey;
     if (event.kind == 9735) {
       for (var tag in event.tags) {
         if (tag[0] == 'bolt11') bolt11 = tag[1];
@@ -37,6 +38,7 @@ class ZapReceipt {
         Nip01Event event = Nip01Event.fromJson(jsonDecode(zapRequestJson));
         comment = event.content;
         sender = event.pubKey;
+        lnurl = event.getFirstTag('lnurl');
         String? amountString = event.getFirstTag('amount');
         if (amountString != null && amountString.isNotEmpty) {
           try {
@@ -54,50 +56,32 @@ class ZapReceipt {
       }
       List<String>? splitStrings = anon?.split('_');
       if (splitStrings != null && splitStrings.length == 2) {
-        // /// recipient decrypt
-        // try {
-        //   String contentBech32 = splitStrings[0];
-        //   String ivBech32 = splitStrings[1];
-        //   String? encryptedContent = bech32Decode(contentBech32,
-        //       maxLength: contentBech32.length)['data'];
-        //   String? iv =
-        //       bech32Decode(ivBech32, maxLength: ivBech32.length)['data'];
-        //
-        //   String encryptedContentBase64 =
-        //       base64Encode(hexToBytes(encryptedContent!));
-        //   String ivBase64 = base64Encode(hexToBytes(iv!));
-        //
-        //   String eventString = await Nip4.decryptContent(
-        //       '$encryptedContentBase64?iv=$ivBase64',
-        //       recipient!,
-        //       myPubkey,
-        //       privkey);
-        //
-        //   /// try to use sender decrypt
-        //   if (eventString.isEmpty) {
-        //     String derivedPrivkey =
-        //         generateKeyPair(recipient, event.createdAt, privkey);
-        //     eventString = await Nip4.decryptContent('$encryptedContent?iv=$iv',
-        //         recipient, bip340.getPublicKey(derivedPrivkey), derivedPrivkey);
-        //   }
-        //   if (eventString.isNotEmpty) {
-        //     Event privEvent = await Event.fromJson(jsonDecode(eventString));
-        //     sender = privEvent.pubkey;
-        //     content = privEvent.content;
-        //   }
-        // } catch (_) {}
+        // TODO decrypt private zap
       }
     } else {
       throw Exception("${event.kind} is not nip57 compatible");
     }
   }
 
-  bool isValid(String invoice) {
-    return bolt11 == invoice;
-    // TODO:
+  /// is valid
+  bool isValid({required String nostrPubKey, required String recipientLnurl}) {
     //  - The zap receipt event's pubkey MUST be the same as the recipient's lnurl provider's nostrPubkey (retrieved in step 1 of the protocol flow).
+    if (pubKey != nostrPubKey) {
+      return false;
+    }
     //  - The invoiceAmount contained in the bolt11 tag of the zap receipt MUST equal the amount tag of the zap request (if present).
+    if (bolt11!=null && bolt11!.isNotEmpty) {
+      if (amountSats != Lnurl.getAmountFromBolt11(bolt11!)) {
+        return false;
+      }
+    }
     //  - The lnurl tag of the zap request (if present) SHOULD equal the recipient's lnurl.
+    if (lnurl!=null && lnurl!.isNotEmpty) {
+      if (lnurl != recipientLnurl) {
+        return false;
+      }
+    }
+    return true;
   }
 
   @override
