+++ packages/ndk/lib/domain_layer/usecases/zaps/zaps.dart
@@ -1,6 +1,9 @@
 import 'dart:async';
+import 'dart:convert';
 
+import 'package:http/http.dart';
 import 'package:ndk/domain_layer/usecases/nwc/nwc_connection.dart';
+import 'package:ndk/domain_layer/usecases/zaps/Invoice_response.dart';
 import 'package:ndk/domain_layer/usecases/zaps/zap_receipt.dart';
 import 'package:ndk/domain_layer/usecases/zaps/zap_request.dart';
 
@@ -14,6 +17,8 @@ import '../nwc/nwc.dart';
 import '../nwc/responses/pay_invoice_response.dart';
 import '../requests/requests.dart';
 
+import 'package:http/http.dart' as http;
+
 /// Zaps
 class Zaps {
   final Requests _requests;
@@ -24,6 +29,94 @@ class Zaps {
     required Nwc nwc,
   })  : _requests = requests,
         _nwc = nwc;
+
+  /// creates an invoice with an optional zap request encoded if signer, pubKey & relays are non empty
+  Future<InvoiceResponse?> fecthInvoice({
+    required String lud16Link,
+    required int amountSats,
+    ZapRequest? zapRequest,
+    String? comment,
+    http.Client? client
+  }) async {
+    var lnurlResponse = await Lnurl.getLnurlResponse(lud16Link, client: client);
+    if (lnurlResponse == null) {
+      return null;
+    }
+
+    var callback = lnurlResponse.callback!;
+    if (callback.contains("?")) {
+      callback += "&";
+    } else {
+      callback += "?";
+    }
+
+    final amount = amountSats * 1000;
+    callback += "amount=$amount";
+
+    if (comment != null && comment.trim() != '') {
+      var commentNum = lnurlResponse.commentAllowed;
+      if (commentNum != null) {
+        if (commentNum < comment.length) {
+          comment = comment.substring(0, commentNum);
+        }
+        callback += "&comment=${Uri.encodeQueryComponent(comment)}";
+      }
+    }
+
+    // ZAP ?
+    if (lnurlResponse.doesAllowsNostr && zapRequest!=null && zapRequest.sig.isNotEmpty) {
+      Logger.log.d(jsonEncode(zapRequest));
+      var eventStr = Uri.encodeQueryComponent(jsonEncode(zapRequest));
+      callback += "&nostr=$eventStr";
+    }
+
+    Logger.log.d("getInvoice callback $callback");
+
+    Uri uri = Uri.parse(callback);
+
+    try {
+      var response = await (client ?? http.Client()).get(uri);
+      final decodedResponse =
+      jsonDecode(utf8.decode(response.bodyBytes)) as Map<String, dynamic>;
+      String invoice = decodedResponse["pr"];
+      return InvoiceResponse(invoice: invoice, amountSats: amountSats, nostrPubkey: lnurlResponse.nostrPubkey);
+
+    } catch (e) {
+      Logger.log.d(e);
+    }
+    return null;
+  }
+
+  Future<ZapRequest> createZapRequest({
+    required int amountSats,
+    required EventSigner signer,
+    required String pubKey,
+    String? eventId,
+    String? comment,
+    required Iterable<String> relays,
+    String? pollOption,
+  }) async {
+    if (amountSats<0) {
+      throw ArgumentError("amount cannot be < 0");
+    }
+    final amount = amountSats * 1000;
+
+    var tags = [
+      ["relays", ...relays],
+      ["amount", amount.toString()],
+      ["p", pubKey],
+    ];
+    if (eventId != null) {
+      tags.add(["e", eventId]);
+    }
+    if (pollOption != null) {
+      tags.add(["poll_option", pollOption]);
+    }
+    var event = ZapRequest(
+        pubKey: signer.getPublicKey(), tags: tags, content: comment??'');
+    await signer.sign(event);
+    return event;
+  }
 
   /// zap or pay some lnurl, for zap to be created it is necessary:
   /// - that the lnurl has the allowsNostr: true
@@ -47,7 +140,7 @@ class Zaps {
         signer != null &&
         relays != null &&
         relays.isNotEmpty) {
-      zapRequest = await Lnurl.zapRequest(
+      zapRequest = await createZapRequest(
           amountSats: amountSats,
           signer: signer,
           pubKey: pubKey,
@@ -55,7 +148,7 @@ class Zaps {
           relays: relays,
           eventId: eventId);
     }
-    String? invoice = await Lnurl.getInvoiceCode(
+    InvoiceResponse? invoice = await fecthInvoice(
       lud16Link: lud16Link!,
       comment: comment,
       amountSats: amountSats,
@@ -66,11 +159,11 @@ class Zaps {
     }
     try {
       PayInvoiceResponse payResponse =
-          await _nwc.payInvoice(nwcConnection, invoice: invoice, timeout: Duration(seconds: 10));
+          await _nwc.payInvoice(nwcConnection, invoice: invoice.invoice, timeout: Duration(seconds: 10));
       if (payResponse.preimage.isNotEmpty && payResponse.errorCode == null) {
         ZapResponse zapResponse = ZapResponse(
             payInvoiceResponse: payResponse);
-        if (zapRequest != null && fetchZapReceipt) {
+        if (zapRequest != null && fetchZapReceipt && invoice.nostrPubkey!=null && invoice.nostrPubkey!.isNotEmpty) {
           // if it's a zap, try to find the zap receipt
           zapResponse.receiptResponse = _requests.subscription(filters: [
             eventId != null
@@ -90,7 +183,7 @@ class Zaps {
             if (bolt11!=null && bolt11 == invoice || preimage!=null && preimage==payResponse.preimage) {
               ZapReceipt receipt = ZapReceipt.fromEvent(event);
               Logger.log.d("Zap Receipt: $receipt");
-              if (receipt.isValid(invoice)) {
+              if (receipt.isValid(nostrPubKey: invoice.nostrPubkey!, recipientLnurl: lnurl)) {
                 zapResponse.emitReceipt(receipt);
               } else {
                 Logger.log.w("Zap Receipt invalid: $receipt");
@@ -120,6 +213,7 @@ class Zaps {
           : Filter(kinds: [ZapReceipt.KIND], pTags: [pubKey])
     ]);
     List<Nip01Event> events = await response.future;
+    // TODO how to check validity of zap receipts without nostrPubKey and recipientLnurl????
     return events.map((event) => ZapReceipt.fromEvent(event)).toList();
   }
